refactor: move preprocess functions to a module.
