+++ src/gtMerger/matchAndMerge.py
@@ -16,15 +16,15 @@
  */
  """
 
-import datetime
 import os
 from pathlib import Path
 
 import pandas as pd
-import pytz
 
 # Import dependencies
 from args import get_parser
+from preprocess import preprocess_gt_data, preprocess_oba_data
+import constants
 
 
 # -------------------------------------------
@@ -50,7 +50,7 @@ def main():
             exit()
 
     # Create sub-folders for output an logs
-    path_logs = os.path.join(command_line_args.dataDir, "logs")
+    path_logs = os.path.join(command_line_args.dataDir, constants.FOLDER_LOGS)
     if not os.path.isdir(path_logs):
         try:
             os.mkdir(path_logs)
@@ -58,7 +58,7 @@ def main():
             print("There was an error while creating the sub folder for logs:", path_logs)
             exit()
 
-    path_output = os.path.join(command_line_args.dataDir, "output")
+    path_output = os.path.join(command_line_args.dataDir, constants.FOLDER_OUTPUT)
     if not os.path.isdir(path_output):
         try:
             os.mkdir(path_output)
@@ -72,113 +72,34 @@ def main():
     gt_data = pd.read_excel(excel_path)
 
     # Preprocess ground truth data
-    gt_data = preprocess_gt_data(gt_data)
+    gt_data, data_gt_dropped = preprocess_gt_data(gt_data)
     print("Ground truth data preprocessed.")
+    # Save data to be dropped to a csv file
+    dropped_file_path = os.path.join(command_line_args.dataDir, constants.FOLDER_LOGS,
+                                     constants.GT_DROPPED_DATA_FILE_NAME)
+    data_gt_dropped.to_csv(path_or_buf=dropped_file_path, index=False)
 
     # Create path OS independent for csv file
     csv_path = Path(command_line_args.obaFile)
     # Load OBA data
     oba_data = pd.read_csv(csv_path)
     # Preprocess OBA data
-    oba_data, data_csv_dropped = preprocess_oba_data(oba_data)
+    oba_data, data_csv_dropped = preprocess_oba_data(oba_data, command_line_args.minActivityDuration,
+                                                     command_line_args.minTripLength)
     print("OBA data preprocessed.")
 
     # Data preprocessing is over
     # Save oba dropped data to a csv file
-    dropped_file_path = os.path.join(command_line_args.dataDir, "logs", "droppedObaData.csv")
+    dropped_file_path = os.path.join(command_line_args.dataDir, constants.FOLDER_LOGS,
+                                     constants.OBA_DROPPED_DATA_FILE_NAME)
     data_csv_dropped.to_csv(path_or_buf=dropped_file_path, index=False)
 
     # merge dataframes
     merged_data_frame = merge(gt_data, oba_data)
 
     # Save merged data to csv
-    merged_file_path = os.path.join(command_line_args.dataDir, "output", "mergedData.csv")
+    merged_file_path = os.path.join(command_line_args.dataDir, constants.FOLDER_OUTPUT, constants.MERGED_DATA_FILE_NAME)
     merged_data_frame.to_csv(path_or_buf=merged_file_path, index=False)
-
-
-def preprocess_oba_data(data_csv) -> object:
-    """ Preprocess the csv data file from oba-firebase-export as follows:
-        - Change activity start date datatype from str to datetime
-        - Drop observations whose activity start date are NaN after data type conversion
-        - Drop observations that does not match the time duration and distance requirements from command_line_args
-          minActivityDuration and minTripLength
-        - Add column required to be used as key while merging with ground truth data
-    
-    Args:
-        data_csv: A data frame loaded from a csv file generated by oba-firebase-export
-
-    Returns:
-        Preprocessed dataframe
-    """
-    # Change Activity Start Date and Time* (UTC) to datetime
-    data_csv['Activity Start Date and Time* (UTC)'] = pd.to_datetime(data_csv['Activity Start Date and Time* (UTC)'],
-                                                                     errors='coerce', utc=True)
-
-    # Drop NaN rows for relevant columns
-    clean_data_csv = data_csv.dropna(
-        subset=['Activity Start Date and Time* (UTC)', 'Origin location Date and Time (*best) (UTC)',
-                'Duration* (minutes)', 'Origin-Destination Bird-Eye Distance* (meters)'])
-
-    # Remove trips with Duration less than command_line_args.minActivityDuration minutes and distance less than
-    # command_line_args.minTripLength
-    clean_data_csv = clean_data_csv[(data_csv['Duration* (minutes)'] >= command_line_args.minActivityDuration) & (
-            data_csv['Origin-Destination Bird-Eye Distance* (meters)'] >= command_line_args.minTripLength)]
-
-    # Add the data to be dropped to a data frame
-    data_csv_dropped = pd.merge(data_csv, clean_data_csv, how='outer', indicator=True).query("_merge != 'both'").drop(
-        '_merge', axis=1).reset_index(drop=True)
-
-    # Return clean data and dropped data as separated dataframes
-    return clean_data_csv, data_csv_dropped
-
-
-def preprocess_gt_data(gt_data):
-    """ Preprocess the ground Truth xlsx data file as follows:
-        - Remove unnamed columns if exist
-        - Change the column to datetime.time
-        - Drop rows with NaN on GT_Date or GT_TimeOrig  after data type conversion
-        - Create GT_DateTimeCombined column joining GT_Date and GT_TimeOrig columns
-        - Assign timezone to GT_DateTimeCombined
-        - Add column required to be used as key while merging with ground truth data
-        - Generates a log with the dropped rows during the preprocess
-    
-    Args:
-        gt_data: A data frame loaded from a xls file generated manually from GT data collection process
-
-    Returns:
-        Preprocessed gt dataframe
-    """
-    # Drop unnamed columns
-    unnamed_cols = [col for col in gt_data.columns if 'Unnamed' in col]
-    gt_data = gt_data.drop(unnamed_cols, axis=1)
-
-    # Change data type of TimeOrig to string to simplify date time conversions
-    gt_data.GT_TimeOrig = gt_data.GT_TimeOrig.astype(str)
-
-    # Change the column to datetime.time, coerce will produce NaN if the change is not possible
-    gt_data['GT_TimeOrig'] = pd.to_datetime(gt_data['GT_TimeOrig'], errors='coerce').dt.time
-
-    # Drop rows with NaN on GT_Date or GT_TimeOrig
-    clean_gt_data = gt_data.dropna(subset=['GT_Date', 'GT_TimeOrig'])
-
-    # Add the data to be dropped to a data frame
-    data_gt_dropped = pd.merge(gt_data, clean_gt_data, how='outer', indicator=True).query("_merge != 'both'").drop(
-        '_merge', axis=1).reset_index(drop=True)
-    # Save data to be dropped to a csv file
-    dropped_file_path = os.path.join(command_line_args.dataDir, "logs", "droppedGtData.csv")
-    data_gt_dropped.to_csv(path_or_buf=dropped_file_path, index=False)
-
-    # Create GT_DateTimeCombined column
-    clean_gt_data.loc[:, 'GT_DateTimeCombined'] = clean_gt_data.apply(
-        lambda x: datetime.datetime.combine(x.GT_Date, x.GT_TimeOrig), 1)
-
-    # Assign timezone to GT_DateTimeCombined
-    clean_gt_data.loc[:, 'GT_DateTimeCombined'] = clean_gt_data.apply(
-        lambda x: pytz.timezone(x.GT_TimeZone).localize(x.GT_DateTimeCombined), 1)
-
-    # Add column to be used in function "merge_asoft"
-    clean_gt_data['ClosestTime'] = clean_gt_data['GT_DateTimeCombined'].dt.tz_convert('UTC')
-    return clean_gt_data
 
 
 def merge(gt_data, oba_data):
